Complete Stage 3 (I'm so lite)
Upgrade the program so it uses a database to store and persist the generated credit cards. Also read the database filename from the command line.

Add an enum class MenuState to control which menu level we're at as well as what choice the user has made. Using these in Main and MenuSystem let us consolidate all the menu handling into one method and switch block in Main.

Write a Database class to handle all the database functions in one place. Takes care of creating the database if it doesn't exist as well as adding and reading records from it.

Main.java:
- add two new instance fields for controlling the state the program is in and holding the credit card information
- consolidate the loggedIn() method with the menu in main()
- add calls to Database methods where needed
- change from using String's to MenuState

MenuSystem.java:
- Change from returning/using String's to using MenuState